Changes how initialization, Topic and predicates are handled.
Moves down extensions to datum instead of obj level.
Makes the extension list lazy initialized.
Predicates split into separate host and user lists, and moved to extension level.
Separates Canuse/Topic() checks into extension_status/act procs.

+++ code/datums/extensions/extensions.dm
@@ -1,55 +1,76 @@
 /datum/extension
-
 	var/datum/holder = null // The holder
+	var/list/host_predicates
+	var/list/user_predicates
+	var/expected_type = /datum
 
-/datum/extension/New(var/datum/holder)
-	if(!istype(holder))
-		CRASH("Invalid holder.")
+/datum/extension/New(var/datum/holder, var/host_predicates = list(), var/user_predicates = list(), var/additional_arguments = list())
+	if(!istype(holder, expected_type))
+		CRASH("Invalid holder type. Expected [expected_type], was [holder.type]")
 	src.holder = holder
 	..()
 
+	src.host_predicates = host_predicates ? host_predicates : list()
+	src.user_predicates = user_predicates ? user_predicates : list()
+
 /datum/extension/Destroy()
 	holder = null
+	host_predicates.Cut()
+	user_predicates.Cut()
 	return ..()
 
-/datum/extension/CanUseTopic(var/mob/user)
-	return holder && user ? STATUS_INTERACTIVE : STATUS_CLOSE
+/datum/extension/proc/extension_status(var/mob/user)
+	if(!holder || !user)
+		return STATUS_CLOSE
+	if(!all_predicates_true(list(holder), host_predicates))
+		return STATUS_CLOSE
+	if(!all_predicates_true(list(user), user_predicates))
+		return STATUS_CLOSE
+	if(holder.CanUseTopic(usr, default_state) != STATUS_INTERACTIVE)
+		return STATUS_CLOSE
 
-/datum/extension/Topic()
+	return STATUS_INTERACTIVE
+
+/datum/extension/proc/extension_act(var/href, var/list/href_list, var/mob/user)
+	return extension_status(user) == STATUS_CLOSE
+
+/datum/extension/Topic(var/href, var/list/href_list)
 	if(..())
-		return 1
-	if(CanUseTopic(usr) != STATUS_INTERACTIVE)
-		return 1
-	return 0
+		return TRUE
+	return extension_act(href, href_list, usr)
 
-/obj
-	var/list/datum/extension/extensions = list()
+/datum
+	var/list/datum/extension/extensions
 
-/obj/Destroy()
-	for(var/expansion_key in extensions)
-		var/list/expansion = extensions[expansion_key]
-		if(islist(expansion))
-			expansion.Cut()
-		else
-			qdel(expansion)
-	extensions.Cut()
+/datum/Destroy()
+	if(extensions)
+		for(var/expansion_key in extensions)
+			var/list/expansion = extensions[expansion_key]
+			if(islist(expansion))
+				expansion.Cut()
+			else
+				qdel(expansion)
+		extensions.Cut()
 	return ..()
 
-/obj/ResetVars(var/list/exclude = list())
+/datum/ResetVars(var/list/exclude = list())
 	exclude += "extensions"
 	..(exclude)
 	//extensions = list()
 
-/proc/set_extension(var/obj/source, var/base_type, var/expansion_type, var/list/arguments = list())
-	source.extensions[base_type] = list(expansion_type, arguments)
+/proc/set_extension(var/datum/source, var/base_type, var/expansion_type, var/host_predicates, var/user_predicates, var/list/additional_argments)
+	if(!source.extensions)
+		source.extensions = list()
+	source.extensions[base_type] = list(expansion_type, host_predicates, user_predicates, additional_argments)
 
-/proc/get_extension(var/obj/source, var/base_type)
+/proc/get_extension(var/datum/source, var/base_type)
+	if(!source.extensions)
+		return
 	var/list/expansion = source.extensions[base_type]
 	if(!expansion)
 		return
 	if(istype(expansion))
 		var/expansion_type = expansion[1]
-		var/expansion_arguments = expansion[2]
-		expansion = new expansion_type(source, expansion_arguments)
+		expansion = new expansion_type(source, expansion[2], expansion[3], expansion[4])
 		source.extensions[base_type] = expansion
 	return expansion
